# Metal: faster Q4_0 and Q4_1 matrix x vector kernels