This PR continues along the lines of #2188 and adds two improvements:
* PR #2188 was merged without the suggested changes that lead to additional 3-5% speed up. This PR adds this additional speedup for `Q4_0`.
* Adding a very similar kernel for matrix vector multiplications for `Q4_1` that leads to a 7-25% speedup compared to master on M2 Max with a 30-core GPU (see table).

Token generation time in milliseconds per token on M2 Max with 30-core GPU using
```
./main -m model -p "I believe the meaning of life is" -c 512 -s 1234 -n 128 -ngl 1 --no-mmap
```

| Model | Q4_0 master | Q4_0 PR | Q4_0 speed up | Q4_1 master | Q4_1 PR | Q4_1 speed up |
|--:|--:|--:|--:|--:|--:|--:|
|7B | 21.1 | 19.8 | 6.6% | 22.1 | 20.7 | 6.8% |
|13B | 35.5 | 33.8 | 5.0% | 39.4 | 34.8 | 13.2% |
|33B | 83.0 | 80.5 | 3.1% | 99.8 | 85.4 | 16.9% |
|65B| 153 | 147 | 4.1% | 194 | 156 | 24.4% |

Some interesting observations:
* The addition check (`if (tiisg == 0 && ((r0 * N_SIMDGROUP + sgitg) * N_DST + row) < ne01)` vs `if (tiisg == 0)` originally) needed to avoid having two separate kernels leads to a measurable performance degradation
* At least on my M2 Max the prefetch of the next quantized block, which was assumed to be the major factor for the #2188 speed up, is a very minor effect. For instance, for `Q4_0` at 7B I measure 19.9 ms/token without prefetch vs 19.8 ms/token with prefetch; at 65B it is 149 ms/token versus 147 ms/token without/with prefetch. Hence, most of the gain in speed comes from letting each thread in a SIMD group process a full block.