@ggerganov 2023-07-13
Results on M1 Pro for `./main -m $model -p "I believe the meaning of life is" --ignore-eos -s 3 -n 128 -t 8 -ngl 1`

| model | master | this PR |
|-|-|-|
| 7B | 35.1 ms/tok | 32.7 ms/tok |
| 13B | 61.2 ms/tok | 57.0 ms/tok |