@ggerganov 2023-07-13
@ikawrakow 

Btw, I don't see the suggested changes in #2188 - did you submit the changes for review or are they still "pending"?

Here is how the PR looks on my end:

<details>

![image](https://github.com/ggerganov/llama.cpp/assets/1991296/7d8d4432-5141-411f-b332-ee471d75225e)

</details>