@ikawrakow 2023-07-13
@ggerganov 
> I just realized something and I am not sure how we haven't spotted this yet:

> In ggml-metal.m we do the following before we dispatch a mul_mat kernel:

This was me throwing out unused arguments to the kernels (at some points I noticed that having fewer arguments passed to the kernels slightly improved performance). But it looks like the corresponding change in `ggml-metal.m` has been lost? In any case, these kernels are never used in matrix-matrix multiplications where it would matter (but it would be useful to fix it anyway).